feat:Added new methods,changed some code parts, UI changed

+++ CompanyApp.Domain/Models/Department.cs
@@ -5,6 +5,6 @@ namespace CompanyApp.Domain.Models
     public class Department:BaseEntity
     {
         public string Name { get; set; }
-        public int Capacity { get; set; }
+        public byte Capacity { get; set; }
     }
 }
